refactor net-table
